requirements.txt
Django==5.1.4
django-bootstrap5==24.3
django-tinymce==4.1.0